Refactor replace testbed test with new test and change README
