--- cmd/jaeger/internal/integration/datareceivers/.nocover
@@ -1,1 +1,0 @@
-A custom testbed data receiver for integration testing purpose
