--- cmd/jaeger/internal/integration/fixtures/generated_pict_pairs_traces.txt
@@ -1,33 +1,0 @@
-Resource	InstrumentationLibrary	Spans
-Empty	None	Several
-K8sCloud	None	One
-VMCloud	None	One
-VMOnPrem	None	All
-VMCloud	None	Several
-Exec	None	One
-VMCloud	Two	One
-VMOnPrem	One	Several
-K8sCloud	One	All
-VMCloud	Two	None
-Faas	None	Several
-K8sCloud	Two	All
-Empty	Two	None
-Empty	One	None
-VMCloud	One	All
-VMOnPrem	None	One
-K8sOnPrem	Two	Several
-Empty	Two	Several
-Exec	Two	Several
-K8sOnPrem	None	One
-K8sCloud	One	None
-Empty	None	All
-Empty	None	One
-VMCloud	None	None
-K8sCloud	Two	One
-Exec	None	None
-K8sOnPrem	One	Several
-VMOnPrem	Two	All
-VMOnPrem	Two	Several
-K8sOnPrem	One	One
-Faas	One	Several
-Faas	None	One
