--- cmd/jaeger/internal/integration/fixtures/grpc_config.yaml
@@ -1,23 +1,0 @@
-service:
-  extensions: [jaeger_storage]
-  pipelines:
-    traces:
-      receivers: [otlp]
-      exporters: [jaeger_storage_exporter]
-
-extensions:
-  jaeger_storage:
-    grpc:
-      some-external-storage:
-        server: localhost:17271
-        connection-timeout: 5s
-
-receivers:
-  otlp:
-    protocols:
-      grpc:
-      http:
-
-exporters:
-  jaeger_storage_exporter:
-    trace_storage: some-external-storage
