--- cmd/jaeger/internal/integration/fixtures/pict_input_traces.txt
@@ -1,3 +1,0 @@
-Resource: Empty, VMOnPrem, VMCloud, K8sOnPrem, K8sCloud, Faas, Exec
-InstrumentationLibrary: None, One, Two
-Spans: None, One, Several, All
