+++ cmd/jaeger/internal/integration/span_writer.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package e2e
+package integration
 
 import (
 	"context"
